feat: add types for `UglifyJS` V3
