+++ types/gulp-filter/tsconfig.json
@@ -14,7 +14,10 @@
         ],
         "types": [],
         "noEmit": true,
-        "forceConsistentCasingInFileNames": true
+        "forceConsistentCasingInFileNames": true,
+        "paths": {
+            "uglify-js": ["uglify-js/v2"]
+        }
     },
     "files": [
         "index.d.ts",
