+++ types/gulp-html-replace/index.d.ts
@@ -2,6 +2,8 @@
 // Project: https://www.npmjs.com/package/gulp-html-replace
 // Definitions by: Peter Juras <https://github.com/peterjuras>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
+// TypeScript Version: 2.2
+
 /// <reference types="node" />
