+++ types/gulp-htmlmin/index.d.ts
@@ -2,6 +2,7 @@
 // Project: https://github.com/jonschlinkert/gulp-htmlmin
 // Definitions by: Tanguy Krotoff <https://github.com/tkrotoff>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
+// TypeScript Version: 2.2
 
 /// <reference types="node" />
