+++ types/gulp-uglify/tsconfig.json
@@ -14,7 +14,12 @@
         ],
         "types": [],
         "noEmit": true,
-        "forceConsistentCasingInFileNames": true
+        "forceConsistentCasingInFileNames": true,
+        "paths": {
+            "uglify-js": [
+                "uglify-js/v2"
+            ]
+        }
     },
     "files": [
         "index.d.ts",
