+++ types/html-minifier/v1/tsconfig.json
@@ -16,6 +16,9 @@
             "html-minifier": [
                 "html-minifier/v1"
             ],
+            "uglify-js": [
+                "uglify-js/v2"
+            ],
             "html-minifier/*": [
                 "html-minifier/v1/*"
             ]
