+++ types/uglify-js/v2/package.json
@@ -1,0 +1,6 @@
+{
+    "private": true,
+    "dependencies": {
+        "source-map": "^0.6.1"
+    }
+}
