+++ types/uglify-js/v2/tsconfig.json
@@ -8,13 +8,16 @@
         "noImplicitThis": true,
         "strictNullChecks": false,
         "strictFunctionTypes": true,
-        "baseUrl": "../",
+        "baseUrl": "../../",
         "typeRoots": [
-            "../"
+            "../../"
         ],
         "types": [],
         "noEmit": true,
-        "forceConsistentCasingInFileNames": true
+        "forceConsistentCasingInFileNames": true,
+        "paths": {
+            "uglify-js": ["uglify-js/v2"]
+        }
     },
     "files": [
         "index.d.ts",
